+++ utils/__init__.py
@@ -1,1 +1,1 @@
-# utils package for Athanor
+# utils package for Citrinitas